+++ internal/provider/types/aws_vpc_peering_gateway_response.go
@@ -14,7 +14,7 @@ type AwsVpcPeeringGatewayResponse struct {
 	ID                             types.String                         `tfsdk:"id"`
 	Name                           types.String                         `tfsdk:"name"`
 	State                          types.String                         `tfsdk:"state"`
-	StateMetadata                  *StateMetadata                       `tfsdk:"state_metadata"`
+	StateMetadata                  *CustomDomainStateMetadata           `tfsdk:"state_metadata"`
 	TransitGatewayAttachmentConfig AwsVpcPeeringGatewayAttachmentConfig `tfsdk:"transit_gateway_attachment_config"`
 	UpdatedAt                      types.String                         `tfsdk:"updated_at"`
 }