+++ internal/sdk/konnect.go
@@ -276,9 +276,9 @@ func WithTimeout(timeout time.Duration) SDKOption {
 // New creates a new instance of the SDK with the provided options
 func New(opts ...SDKOption) *Konnect {
 	sdk := &Konnect{
-		SDKVersion: "2.9.0",
+		SDKVersion: "2.10.0",
 		sdkConfiguration: config.SDKConfiguration{
-			UserAgent:  "speakeasy-sdk/terraform 2.9.0 2.620.2 2.0.0 github.com/kong/terraform-provider-konnect/v2/internal/sdk",
+			UserAgent:  "speakeasy-sdk/terraform 2.10.0 2.620.2 2.0.0 github.com/kong/terraform-provider-konnect/v2/internal/sdk",
 			ServerList: ServerList,
 		},
 		hooks: hooks.New(),